move sampleBy to stat
